added CSRF Token for user-list.php and delete-account-confirmation.php while removing useless console.log on characterCreator.js create_character()

+++ page/error/delete-account-confirmation.php
@@ -1,5 +1,5 @@
 <?php 
-	if(isset($_POST["confirm"])){
+	if(isset($_POST["confirm"]) && check_csrf_token()){
 		delete_self_account();
 	}
 ?>
@@ -12,6 +12,7 @@
 <body>
 	<?php require_once('layout/header/header.php'); ?>
     <div class="margin-12">
+		<?php embed_csrf_token() ?>
         <H1>Vous êtes sur le point de supprimer votre compte et toutes vos données de façon permanente et ceci ne peut être annulé une fois cliqué sur le bouton de suppression!</H1><br>
 		<h2>Souhaitez-vous vraiment procéder à la suppression?</h2>
      	<button class="color-red vw-075" onclick="deleteSelfAccount()"><h2>Confirmer la suppression permanente de mon compte</h2></button> <button class="color-secondary-darker vw-075" onclick="location.href='/menu'";><h2>Annuler la suppression de mon compte</h2></button>
